Present helper controller on UI Tests
