+++ Wikipedia/Code/ExploreViewController.swift
@@ -62,7 +62,9 @@ class ExploreViewController: ColumnarCollectionViewController, ExploreCardViewCo
         detailTransitionSourceRect = nil
         logFeedImpressionAfterDelay()
         dataStore.remoteNotificationsController.loadNotifications(force: false)
-        presentEditorForUITest()
+        #if UITEST
+        presentUITestHelperController()
+        #endif
     }
     
     override func viewWillHaveFirstAppearance(_ animated: Bool) {
@@ -79,89 +81,9 @@ class ExploreViewController: ColumnarCollectionViewController, ExploreCardViewCo
         (parent as? WMFAppViewController)?.navigationItem.backButtonTitle = title
     }
 
-
-    func presentEditorForUITest() {
-
-        let localizedStrings = WKSourceEditorLocalizedStrings(inputViewTextFormatting: CommonStrings.textFormatting,
-                                                              inputViewStyle: CommonStrings.style,
-                                                              inputViewClearFormatting: CommonStrings.clearFormatting,
-                                                              inputViewParagraph: CommonStrings.paragraph,
-                                                              inputViewHeading: CommonStrings.heading,
-                                                              inputViewSubheading1: CommonStrings.subheading1,
-                                                              inputViewSubheading2: CommonStrings.subheading2,
-                                                              inputViewSubheading3: CommonStrings.subheading3,
-                                                              inputViewSubheading4: CommonStrings.subheading4,
-                                                              findReplaceTypeSingle: CommonStrings.findAndReplaceSingle,
-                                                              findReplaceTypeAll: CommonStrings.findAndReplaceAll,
-                                                              findReplaceWith: CommonStrings.replaceWith,
-                                                              accessibilityLabelButtonFormatText: CommonStrings.accessibilityLabelButtonFormatText,
-                                                              accessibilityLabelButtonFormatHeading: CommonStrings.accessibilityLabelButtonFormatHeading,
-                                                              accessibilityLabelButtonCitation: CommonStrings.accessibilityLabelButtonCitation,
-                                                              accessibilityLabelButtonCitationSelected: CommonStrings.accessibilityLabelButtonCitationSelected,
-                                                              accessibilityLabelButtonLink: CommonStrings.accessibilityLabelButtonBold,
-                                                              accessibilityLabelButtonLinkSelected: CommonStrings.accessibilityLabelButtonLinkSelected,
-                                                              accessibilityLabelButtonTemplate: CommonStrings.accessibilityLabelButtonTemplate,
-                                                              accessibilityLabelButtonTemplateSelected: CommonStrings.accessibilityLabelButtonTemplateSelected,
-                                                              accessibilityLabelButtonMedia: CommonStrings.accessibilityLabelButtonMedia,
-                                                              accessibilityLabelButtonFind: CommonStrings.accessibilityLabelButtonFind,
-                                                              accessibilityLabelButtonListUnordered: CommonStrings.accessibilityLabelButtonListUnordered,
-                                                              accessibilityLabelButtonListUnorderedSelected: CommonStrings.accessibilityLabelButtonListUnorderedSelected,
-                                                              accessibilityLabelButtonListOrdered: CommonStrings.accessibilityLabelButtonListOrdered,
-                                                              accessibilityLabelButtonListOrderedSelected: CommonStrings.accessibilityLabelButtonListOrderedSelected,
-                                                              accessibilityLabelButtonInceaseIndent: CommonStrings.accessibilityLabelButtonIncreaseIndent,
-                                                              accessibilityLabelButtonDecreaseIndent: CommonStrings.accessibilityLabelButtonDecreaseIndent,
-                                                              accessibilityLabelButtonCursorUp: CommonStrings.accessibilityLabelButtonCursorUp,
-                                                              accessibilityLabelButtonCursorDown: CommonStrings.accessibilityLabelButtonCursorDown,
-                                                              accessibilityLabelButtonCursorLeft: CommonStrings.accessibilityLabelButtonCursorLeft,
-                                                              accessibilityLabelButtonCursorRight: CommonStrings.accessibilityLabelButtonCursorRight,
-                                                              accessibilityLabelButtonBold: CommonStrings.accessibilityLabelButtonBold,
-                                                              accessibilityLabelButtonBoldSelected: CommonStrings.accessibilityLabelButtonBoldSelected,
-                                                              accessibilityLabelButtonItalics: CommonStrings.accessibilityLabelButtonItalics,
-                                                              accessibilityLabelButtonItalicsSelected: CommonStrings.accessibilityLabelButtonItalicsSelected,
-                                                              accessibilityLabelButtonClearFormatting: CommonStrings.accessibilityLabelButtonClearFormatting,
-                                                              accessibilityLabelButtonShowMore: CommonStrings.accessibilityLabelButtonShowMore,
-                                                              accessibilityLabelButtonComment: CommonStrings.accessibilityLabelButtonComment,
-                                                              accessibilityLabelButtonCommentSelected: CommonStrings.accessibilityLabelButtonCommentSelected,
-                                                              accessibilityLabelButtonSuperscript: CommonStrings.accessibilityLabelButtonSuperscript,
-                                                              accessibilityLabelButtonSuperscriptSelected: CommonStrings.accessibilityLabelButtonSuperscriptSelected,
-                                                              accessibilityLabelButtonSubscript: CommonStrings.accessibilityLabelButtonSubscript,
-                                                              accessibilityLabelButtonSubscriptSelected: CommonStrings.accessibilityLabelButtonSubscriptSelected,
-                                                              accessibilityLabelButtonUnderline: CommonStrings.accessibilityLabelButtonUnderline,
-                                                              accessibilityLabelButtonUnderlineSelected: CommonStrings.accessibilityLabelButtonUnderlineSelected,
-                                                              accessibilityLabelButtonStrikethrough: CommonStrings.accessibilityLabelButtonStrikethrough,
-                                                              accessibilityLabelButtonStrikethroughSelected: CommonStrings.accessibilityLabelButtonStrikethroughSelected,
-                                                              accessibilityLabelButtonCloseMainInputView: CommonStrings.accessibilityLabelButtonCloseMainInputView,
-                                                              accessibilityLabelButtonCloseHeaderSelectInputView: CommonStrings.accessibilityLabelButtonCloseHeaderSelectInputView,
-                                                              accessibilityLabelFindTextField: CommonStrings.accessibilityLabelButtonFind,
-                                                              accessibilityLabelFindButtonClear: CommonStrings.accessibilityLabelButtonClearFormatting,
-                                                              accessibilityLabelFindButtonClose: CommonStrings.accessibilityLabelFindButtonClose,
-                                                              accessibilityLabelFindButtonNext: CommonStrings.accessibilityLabelFindButtonNext,
-                                                              accessibilityLabelFindButtonPrevious: CommonStrings.accessibilityLabelFindButtonPrevious,
-                                                              accessibilityLabelReplaceTextField: CommonStrings.accessibilityLabelReplaceTextField,
-                                                              accessibilityLabelReplaceButtonClear: CommonStrings.accessibilityLabelReplaceButtonClear,
-                                                              accessibilityLabelReplaceButtonPerformFormat: CommonStrings.accessibilityLabelReplaceButtonPerformFormat,
-                                                              accessibilityLabelReplaceButtonSwitchFormat: CommonStrings.accessibilityLabelReplaceButtonSwitchFormat,
-                                                              accessibilityLabelReplaceTypeSingle: CommonStrings.accessibilityLabelReplaceTypeSingle,
-                                                              accessibilityLabelReplaceTypeAll: CommonStrings.accessibilityLabelReplaceTypeAll)
-
-        let accessibilityId = WKSourceEditorAccessibilityIdentifiers(
-            textView: SourceEditorAccessibilityIdentifiers.textView.rawValue,
-            findButton: SourceEditorAccessibilityIdentifiers.findButton.rawValue,
-            showMoreButton: SourceEditorAccessibilityIdentifiers.showMoreButton.rawValue,
-            closeButton: SourceEditorAccessibilityIdentifiers.closeButton.rawValue,
-            formatTextButton: SourceEditorAccessibilityIdentifiers.formatTextButton.rawValue,
-            formatHeadingButton: SourceEditorAccessibilityIdentifiers.formatHeadingButton.rawValue,
-            expandingToolbar: SourceEditorAccessibilityIdentifiers.expandingToolbar.rawValue,
-            highlightToolbar: SourceEditorAccessibilityIdentifiers.highlightToolbar.rawValue,
-            findToolbar: SourceEditorAccessibilityIdentifiers.findButton.rawValue,
-            mainInputView: SourceEditorAccessibilityIdentifiers.mainInputView.rawValue,
-            headerSelectInputView: SourceEditorAccessibilityIdentifiers.headerSelectInputView.rawValue
-          )
-
-        let viewModel = WKSourceEditorViewModel(configuration: .full, initialText: "UITest", accessibilityIdentifiers: accessibilityId, localizedStrings: localizedStrings)
-        let editor = WKSourceEditorViewController(viewModel: viewModel, delegate: self)
-
-        present(editor, animated: false)
+    func presentUITestHelperController() {
+        let viewController = UITestHelperViewController(theme: theme)
+        present(viewController, animated: false)
     }
 
     private func restoreScrollPositionIfNeeded() {
